fix: corregir tipo de id en Editorial, Idioma y EstadoLibro de Long a Integer para coincidir con SERIAL del esquema

+++ backend-springboot/src/main/java/com/uteq/backend/repository/EstadoLibroRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EstadoLibroRepository extends JpaRepository<EstadoLibro, Long> {}
+public interface EstadoLibroRepository extends JpaRepository<EstadoLibro, Integer> {}
